Refactoring Error Handling.
This is going in the final direction now: given the detailed requirements that the WireProtocol
has in terms of Error Handling (http://code.google.com/p/selenium/wiki/JsonWireProtocol#Error_Handling).
This commit covers the "Invalid Request Errors".

The idea I have in mind is to have every single WireProtocol-complieant error,
able to carry it's own error handling, and the format of what is returned.