Workaround to make Python WebDriver binding happy.
When a request `"Content-Type"` equals `"application/x-www-form-urlencoded"`,
PhantomJS assumes (correctly?) that the POST Body Message is a set of values
if a form URL-Encoded format.
So it parses the content of the body, creating a QVariantMap (JSON),
that is then attached to the `request.post` property.

The "as-is" content is stored in `request.postRaw`.

This workaround overrides `request.post = request.postRaw`,
but the issue should be solved on the Python Binding level that shouldn't
set the wrong "Content-Type".

UPDATE: Based on a conversation with David Burns, the Python bindings in
Selenium 2.26 have that fixed.
See [Selenium Issue #4630](http://code.google.com/p/selenium/issues/detail?id=4630).

Fixes Issue #6.