Unify `execFuncAndWaitForLoad` + cleared callbacks
The presence of both `execFuncAndWaitForLoad` and
`evaluateAndWaitForLoad` was starting to irritate
me: two methods, almost identical. They needed
to be unified.
And that's what I did.

Also, I'm making sure that at the end of whatever
callback (or timeout) is invoked, all the registered
`oneShotCallback` that were registered but never
cleared because never invoked (DAH!) are now…
cleared.