FIX frame switching for when using ELEMENT ID.
In WebDriver you can switch to a Frame via it's
WebElement reference. That looks like this: `{ "ELEMENT" : "903209u209…" }`.

GhostDriver was able to handle that, as long as the frame had a `name` attribute.
It was not able to deal with frame that didn't have that, but just `id` attribute.

This fixes those scenarios.

Fixes #159.