Create the first Window of a Session on creation.
I was trying to be "smart" and delay the creation
of the actual Page/Window to the moment in which
the first interaction with it would actually
happen.

Unfortunately this clashes in a race condition
with the timer that takes care of deleting sessions
that have no more windows/pages: a newly created
session would be "windowless" until the first
interaction, and that would cause the issue.

Fixes issue #172.