Introduced "config.js" to manage runtime config.
We need to add few command line parameters to
control Logging and so I took the time to do it
in a more organic way.

This also means that the Command Line params
that GhostDriver requires are going to change
substantially. The effect of this "compatibility breakage"
shouldn't be seen by PhantomJS users (the "normal"
users) as the CLI there won't change.
I'll just have to add NEW parameters for the Logging.