Handle "thisWindow.close()" cases.
It's possible that some windows contain javascript
code to close the very window we are pointing at.
This was causing GhostDriver to hang, as it was
assuming "THE" page to which that window was associated
would be always available.

The fix is pretty easy: we just ignore the fact
that the window is now gone and continue happily.

Fixes #180.