Session Isolation within the same Process instance.
Finally an API to support multi-cookie jar is available in PhantomJS:
https://github.com/ariya/phantomjs/pull/11535.

This allows us to create a completely new CookieJar every time
we create a Session.

A version of PhantomJS with the new CookieJar API
hasn't been released yet: once the commit linked above
is part of a stable release, this will work.
Otherwise, stick with GhostDriver 1.1.0.

Fixes #170.